+++ src/main/java/com/ma_recruit/repository/community/CommunityPostRepository.java
@@ -3,7 +3,5 @@ package com.ma_recruit.repository.community;
 import com.ma_recruit.entity.community.CommunityPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface CommunityPostRepository extends JpaRepository<CommunityPost, BigInteger> {
+public interface CommunityPostRepository extends JpaRepository<CommunityPost, Integer> {
 }
