+++ src/main/java/com/ma_recruit/repository/member/ProfileRepository.java
@@ -3,7 +3,5 @@ package com.ma_recruit.repository.member;
 import com.ma_recruit.entity.member.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface ProfileRepository extends JpaRepository<Profile, BigInteger> {
+public interface ProfileRepository extends JpaRepository<Profile, Integer> {
 }
