+++ src/main/java/com/ma_recruit/repository/party/MapleMapRepository.java
@@ -3,7 +3,5 @@ package com.ma_recruit.repository.party;
 import com.ma_recruit.entity.party.MapleMap;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface MapleMapRepository extends JpaRepository<MapleMap, BigInteger> {
+public interface MapleMapRepository extends JpaRepository<MapleMap, Integer> {
 }
