+++ src/main/java/com/ma_recruit/repository/party/WorldMapRepository.java
@@ -3,7 +3,5 @@ package com.ma_recruit.repository.party;
 import com.ma_recruit.entity.party.WorldMap;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface WorldMapRepository extends JpaRepository<WorldMap, BigInteger> {
+public interface WorldMapRepository extends JpaRepository<WorldMap, Integer> {
 }
